remove all old annotations code
