+++ client/browser-ext/app/utils/annotations.js
@@ -4,176 +4,258 @@ import styles from "../components/App.css";
 import _ from "lodash";
 import EventLogger from "../analytics/EventLogger";
 
-// addAnnotations takes json annotation data returned from the
-// Sourcegraph annotations API and manipulates the DOM to add
-// hover-over tooltips and links.
-//
-// It assumes the caller has verified that the current view
-// is "ready" to be annotated (e.g. DOM elements have all been rendered)
-// and that there are no overlapping annotations in the json
-// returned by the Sourcegraph API.
-//
-// It assumes that the formatted html provided by the Sourcegraph API
-// for doc tooltips is "safe" to be injected into the page.
-//
-// It does *not* assume that the code that is being annotated is safe
-// to be executed in our script, so we take care to properly escape
-// characters during the annotation loop.
-export default function addAnnotations(json) {
-	if (document.getElementById("sourcegraph-annotation-marker")) {
-		// This function is not idempotent; don't let it run twice.
+// addAnnotations is the entry point for injecting annotations onto a blob (el).
+// An invisible marker is appended to the document to indicate that annotation
+// has been completed; so this function expects that it will be called once all
+// repo/annotation data is resolved from the server.
+export default function addAnnotations(path, el, anns, lineStartBytes) {
+	const markerID = `sourcegraph-annotation-marker-${path}`;
+	if (document.getElementById(markerID)) {
+		// Don't let annotations be applied twice (it is not idempotent).
 		return;
+	} else {
+		const marker = document.createElement("div");
+		marker.id = markerID;
+		marker.style.display = "none";
+		el.appendChild(marker);
 	}
-
-	let annsByStartByte = {};
-	let annsByEndByte = {};
-	for (let i = 0; i < json.Annotations.length; i++){
-		if (json.Annotations[i].URL) {
-			let ann = json.Annotations[i];
-			annsByStartByte[ann.StartByte] = ann;
-			annsByEndByte[ann.EndByte] = ann;
-		}
-	}
-	traverseDOM(annsByStartByte, annsByEndByte);
-
-	// Prevent double annotation on any file by adding some hidden
-	// state to the page.
-	const el = document.querySelector(".blob-wrapper");
-	if (el) {
-		const annotationMarker = document.createElement("div");
-		annotationMarker.id = "sourcegraph-annotation-marker";
-		annotationMarker.style.display = "none";
-		el.appendChild(annotationMarker);
-	}
+	_applyAnnotations(el, indexAnnotations(anns).annsByStartByte, indexLineStartBytes(lineStartBytes));
+	_postProcess();
 }
 
-let annotating = false; // imperative private value indicating whether annotation is in progress for a single token (def)
+let _annURLs = []; // global state container for annotation URLs matched to tokens
+// _postProcess fetches metadata for each of the repositories matched to the
+// page within annotations
+function _postProcess() {
+	_annURLs.forEach((url) => setTimeout(() => cacheDefaultBranch(url)));
+	_annURLs = [];
+}
+// _applyAnnotations is a helper function for addAnnotations
+export function _applyAnnotations(el, annsByStartByte, startBytesByLine) {
+	// The blob is represented by a table; the first column is the line number,
+	// the second is code. Each row is a line of code
+	const table = el.querySelector("table");
 
-// traverseDOM handles the actual DOM manipulation.
-function traverseDOM(annsByStartByte, annsByEndByte){
-	let table = document.querySelector("table");
-	let count = 0;
+	for (let i = 0; i < table.rows.length; ++i) {
+		const row = table.rows[i];
+		const line = row.cells[0].dataset.lineNumber;
 
-	// get output HTML for each line and replace the original <td>
-	for (let i = 0; i < table.rows.length; i++){
-		let output = "";
-		let row = table.rows[i];
+		const codeCell = row.cells[1];
+		const offset = startBytesByLine[line];
 
-		// Code is always the second <td> element; we want to replace code.innerhtml
-		// with a Sourcegraph-"linkified" version of the token, or the same token
-		let code = row.cells[1]
-		let children = code.childNodes;
-		let startByte = count;
-		count += utf8.encode(code.textContent).length;
-		if (code.textContent !== "\n") {
-			count++; // newline
-		}
-
-		for (let j = 0; j < children.length; j++) {
-			let childNodeChars; // the "inner-stuff" of the code cell
-
-			if (children[j].nodeType === Node.TEXT_NODE){
-				childNodeChars = children[j].nodeValue.split("");
-			} else {
-				// Quotation marks on GitHub are given their own span tags.
-				// This messes up anns for go imports and other strings with quote marks.
-				// We handle this by combining all the child span tags when we see child
-				// spans with "pl-pds", the class for quotes, and there aren't any other
-				// tags between them (tags in between needed for highlighting).
-				if (children[j].children.length > 1) {
-					let stringWithQuotes = true;
-					for (let k = 0; k < children[j].children.length; k++) {
-						if (children[j].children[k].className !== "pl-pds") {
-							stringWithQuotes = false
-							break;
-						}
-					}
-					if (stringWithQuotes) {
-						children[j].innerHTML = _.escape(children[j].innerText);
-					}
-				}
-				childNodeChars = children[j].outerHTML.split("");
-			}
-
-			let consumingSpan = false;
-			annotating = false;
-
-			// go through each char of childNodes
-			for (let k = 0; k < childNodeChars.length; k++) {
-				if (childNodeChars[k] === "<" && (childNodeChars.slice(k, k+5).join("") === "<span" || childNodeChars.slice(k, k+6).join("") === "</span")) {
-					consumingSpan = true;
-				}
-				if (!consumingSpan){
-					// Case to handle if < or > appears, so that we don't "consume" or make span tags in the code disappear
-					// This will not break if "&lt;" or "&gt;" appear because chars are escaped.
-					if (isStartOfSpanTag(childNodeChars, k)) {
-						output += next(childNodeChars.slice(k, k+4).join(""), startByte, annsByStartByte, annsByEndByte);
-						k += childNodeChars.slice(k, k+4).join("").length-1;
-						startByte += utf8.encode(childNodeChars[k]).length;
-					}
-					else {
-						output += next(childNodeChars[k], startByte, annsByStartByte, annsByEndByte);
-						startByte += utf8.encode(childNodeChars[k]).length;
-					}
-				} else {
-					// when we are consuming the <span> element, don't increment startByte
-					output += childNodeChars[k];
-				}
-
-				if (childNodeChars[k] === ">" && consumingSpan) {
-					consumingSpan = false;
-				}
-			}
-		}
+		// result is the new (annotated) innerHTML of the code cell
+		const {result, bytesConsumed} = convertNode(codeCell, annsByStartByte, offset);
 
 		// manipulate the DOM asynchronously so the page doesn't freeze while large
 		// code files are being annotated
 		setTimeout(() => {
-			code.innerHTML = output;
-			let newRows = code.childNodes
-			for (let n = 0; n < newRows.length; n++) {
-				addPopover(newRows[n]);
-			}
+			codeCell.innerHTML = result;
+			addPopover(codeCell);
 		});
 	}
 }
 
-function isStartOfSpanTag(childNodeChars, idx) {
-	return (childNodeChars[idx] === "&" && (((childNodeChars.slice(idx, idx+4).join("")) === ("&gt;")) || (childNodeChars.slice(idx, idx+4).join("") === ("&lt;"))))
-}
-
-// next is a helper method for traverseDOM which transforms a character
-// into itself or wraps the character in a starting/ending anchor tag
-function next(c, byteCount, annsByStartByte, annsByEndByte) {
-	if (c === "<" || c === ">") {
-		c = _.escape(c);
-	}
-	let matchDetails = annsByStartByte[byteCount];
-	// if there is a match
-	if (!annotating && matchDetails) {
-		// Handle non-GitHub defs by going to Sourcegraph.
-		const defIsOnGitHub = matchDetails.URL && matchDetails.URL.includes("github.com/");
-
-		const url = defIsOnGitHub ? urlToDef(matchDetails.URL) : `https://sourcegraph.com${matchDetails.URL}`;
-		if (!url) return c;
-
-		const insert = `<a href="${url}" ${defIsOnGitHub ? "data-sourcegraph-ref" : "target=tab"} data-src="https://sourcegraph.com${matchDetails.URL}" class=${styles.sgdef}>${c}`;
-
-		// off-by-one case
-		if (annsByStartByte[byteCount].EndByte - annsByStartByte[byteCount].StartByte === 1) {
-			return `${insert}</a>`;
+// indexAnnotations creates a fast lookup structure optimized to query
+// annotations by start or end byte.
+export function indexAnnotations(anns) {
+	let annsByStartByte = {};
+	let annsByEndByte = {};
+	for (let i = 0; i < anns.length; i++) {
+		if (anns[i].URL) { // without a URL, it is a syntax highlighting annotation
+			let ann = anns[i];
+			annsByStartByte[ann.StartByte] = ann;
+			annsByEndByte[ann.EndByte] = ann;
 		}
-		cacheDefaultBranch(matchDetails.URL)
-		annotating = true;
-		return insert;
 	}
-	// if we reach the end, close the tag.
-	if (annotating && annsByEndByte[byteCount + 1]) {
-		annotating = false;
-		return `${c}</a>`;
-	}
-	return c;
+	return {annsByStartByte, annsByEndByte};
 }
+
+// indexLineStartBytes creates a fast lookup structure optimized to query
+// byte offsets by line number (1-indexed).
+export function indexLineStartBytes(lineStartBytes) {
+	let startBytesByLine = {};
+	for (let i = 0; i < lineStartBytes.length; i++) {
+		startBytesByLine[i + 1] = lineStartBytes[i];
+	}
+	return startBytesByLine;
+}
+
+// annGenerator returns a "match" object if an anotation is defined
+// at the byte offset. The match result contains the number of bytes
+// matched by the annotation, and a generator function which returns
+// an HTML anchor tag string .
+export function annGenerator(annsByStartByte, byte) {
+	const match = annsByStartByte[byte];
+	if (!match) return null;
+
+	function urlToDef(origURL) {
+		if (!origURL) return null;
+		const parts = origURL.split("/-/");
+		if (parts.length < 2) return null;
+		const repo = parts[0]; // remove leading slash
+		const def = parts.slice(1).join("/-/").replace("def/", "");
+		if (repo.startsWith("/github.com/")) {
+			return `https:/${repo}#sourcegraph&def=${def}`;
+		}
+		return `https://github.com/#sourcegraph&repo=${repo}&def=${def}`;
+	}
+
+	const defIsOnGitHub = match.URL && match.URL.includes("github.com/");
+	const url = defIsOnGitHub ? urlToDef(match.URL) : `https://sourcegraph.com${match.URL}`;
+
+	const annLen = match.EndByte - match.StartByte;
+	return {annLen, annGen: function(innerHTML) {
+		_annURLs.push(url);
+		return `<a href="${url}" ${defIsOnGitHub ? "data-sourcegraph-ref" : "target=tab"} data-src="https://sourcegraph.com${match.URL}" class=${styles.sgdef}>${innerHTML}</a>`;
+	}};
+}
+
+// getOpeningTag returns the starting tag (with attributes) of the node
+// (assumed to be of type NodeType.ELEMENT_NODE). E.g.
+//     <span attr="foo">hello world</span>
+// would return '<span attr="foo">'.
+// This is a fairly naive implementation that may not work if we were writing a
+// full-blown HTML parser; but since we only have to parse GitHub's blob HTML
+// we can expect more regularity.
+export function getOpeningTag(node) {
+	let i;
+	let inAttribute = false;
+	const outerHTML = node.outerHTML;
+	for (i = 0; i < outerHTML.length; ++i) {
+		if (outerHTML[i] === "\"") inAttribute = !inAttribute;
+		if (outerHTML[i] === ">" && !inAttribute) break;
+	}
+	return outerHTML.substring(0, i+1);
+}
+
+// convertNode takes a DOM node and returns an object containing the
+// maybe-linkified version of the node as an HTML string as well as the number of bytes consumed.
+// It is the entry point for converting a <td> "cell" representing a line of code.
+// It may also be called recursively for children (which are assumed to be <span>
+// code highlighting annotations from GitHub).
+export function convertNode(node, annsByStartByte, offset) {
+	let result, bytesConsumed, c; // c is an intermediate result
+	if (node.nodeType === Node.ELEMENT_NODE) {
+		// The logic here is to:
+		//    - convert as element node (which may be the special-cased "quoted string" node)
+		//    - ^^ gives inner html; wrap this with the node's current syntax highlighting <span>
+		//    - ^^ but don't do that if the top-level tag is the <td> element (entrypoint)
+
+		c = isQuotedStringNode(node) ?
+			convertQuotedStringNode(node, annsByStartByte, offset) :
+			convertElementNode(node, annsByStartByte, offset);
+
+		if (node.tagName !== "TD") {
+			const openTag = getOpeningTag(node);
+			const closeTag = "</span>";
+			if (openTag.indexOf("<span") !== 0) {
+				throw new Error(`element node tag is not SPAN, got(${node.tagName}), parsed(${openTag})`);
+			}
+			result = `${openTag}${c.result}${closeTag}`;
+		} else {
+			result = c.result;
+		}
+		bytesConsumed = c.bytesConsumed;
+	} else if (node.nodeType === Node.TEXT_NODE) {
+		c = convertTextNode(node, annsByStartByte, offset);
+		result = c.result;
+		bytesConsumed = c.bytesConsumed;
+	} else {
+		throw new Error(`unexpected node type(${node.nodeType})`);
+	}
+
+	return {result, bytesConsumed};
+}
+
+// convertTextNode takes a DOM node which should be of NodeType.TEXT_NODE
+// (this must be checked by the caller) and returns an object containing the
+//  maybe-linkified version of the node as an HTML string and the number
+// of bytes consumed.
+export function convertTextNode(node, annsByStartByte, offset) {
+	let innerHTML = [];
+	let bytesConsumed;
+
+	// Text could contain escaped character sequences (e.g. "&gt;") or UTF-8
+	// decoded characters (e.g. "˟") which need to be properly counted in terms of bytes.
+	const nodeText = utf8.encode(_.unescape(node.wholeText)).split("");
+	for (bytesConsumed = 0; bytesConsumed < nodeText.length;) {
+		const match = annGenerator(annsByStartByte, offset + bytesConsumed);
+		if (!match) {
+			innerHTML.push(_.escape(nodeText[bytesConsumed++]));
+			continue;
+		}
+
+		innerHTML.push(match.annGen(_.escape(nodeText.slice(bytesConsumed, bytesConsumed + match.annLen).join(""))));
+		bytesConsumed += match.annLen;
+	}
+
+	return {result: utf8.decode(innerHTML.join("")), bytesConsumed};
+}
+
+// convertElementNode takes a DOM node which should be of NodeType.ELEMENT_NODE
+// (this must be checked by the caller) and returns an object containing the
+//  maybe-linkified version of the node as an HTML string as well as the number of bytes consumed.
+export function convertElementNode(node, annsByStartByte, offset) {
+	let innerHTML = [];
+	let bytesConsumed = 0;
+
+	// The logic here is to simply recurse on each of the child nodes; everything is eventually
+	// just a text node or the special-cased "quoted string node" (see below).
+	for (let i = 0; i < node.childNodes.length; ++i) {
+		const res = convertNode(node.childNodes[i], annsByStartByte, offset + bytesConsumed);
+		innerHTML.push(res.result);
+		bytesConsumed += res.bytesConsumed;
+	}
+
+	return {result: utf8.decode(innerHTML.join("")), bytesConsumed};
+}
+
+// isQuotedStringNode is a predicate function to identify the special-cased
+// string identifier DOM element, which takes a like of go code like this:
+//
+//    import (
+//        "fmt"
+//    )
+//
+// and creates for the "fmt" and creates this:
+//
+//    "<span class="pl-s"><span class="pl-pds">"</span>fmt<span class="pl-pds">"</span></span>"
+//
+// Without the special-casing the <a class="sg-link" /> tag will be put around the opening quote,
+// but the total # of bytes consumed would automatically count the rest of the fmt".
+// This guarantees the annotation consumes the entire set of childNodes.
+export function isQuotedStringNode(node) {
+	return node.childNodes.length === 3 && node.querySelectorAll(".pl-pds").length === 2 &&
+		node.innerText.startsWith("\"") && node.innerText.endsWith("\"");
+}
+
+// convertQuotedStringNode takes a DOM node which should pass the isQuotedStringNode predicate
+// (this must be checked by the caller) and returns an object containing the
+//  maybe-linkified version of the node as an HTML string as well as the number of bytes consumed.
+export function convertQuotedStringNode(node, annsByStartByte, offset) {
+	const text = `"${utf8.encode(_.unescape(node.childNodes[1].wholeText))}"`; // put quotes around the sanitized inner text
+	const match = annGenerator(annsByStartByte, offset);
+
+	// NOTE:
+	// match could be undefined if the annotation doesn't consume the opening start quote;
+	// we assume there is no chance that the string inside the quotes would otherwise have annotations.
+	// ^^ this assumption may break some day, but I haven't seen it do so yet. Even so, we could
+	// check/theoretically handle this case, but I think it is better not to add more special casing to
+	// the special casing which is not yet observed in the wild.
+
+	if (!match) return {result: node.innerHTML, bytesConsumed: text.length};
+	if (match.annLen !== text.length) {
+		throw new Error(`annotation for quoted string node has length mismatch, got(${match.annLen}) wanted(${text.length})`);
+	}
+	return {result: match.annGen(node.innerHTML), bytesConsumed: match.annLen};
+}
+
+
+
+// The rest of this file is responsible for fetching/caching annotation specific data from the server
+// and adding interaction popover data to annotated elements.
+// The sate management is done outside of the Redux container, thought it could be there; some of this
+// stuff we don't need synchonized to browser local storage.
 
 export const defaultBranchCache = {};
 // fetchingDefaultBranchCache ensures we only make one API call per repo to get default branch.
@@ -195,18 +277,6 @@ function cacheDefaultBranch(annURL) {
 			})
 			.catch((err) => console.log("Error getting default branch"))
 	}
-}
-
-function urlToDef(origURL) {
-	if (!origURL) return null;
-	const parts = origURL.split("/-/");
-	if (parts.length < 2) return null;
-	const repo = parts[0]; // remove leading slash
-	const def = parts.slice(1).join("/-/").replace("def/", "");
-	if (repo.startsWith("/github.com/")) {
-		return `https:/${repo}#sourcegraph&def=${def}`;
-	}
-	return `https://github.com/#sourcegraph&repo=${repo}&def=${def}`;
 }
 
 let popoverCache = {};
